Dot_AI/backend/requirements.txt
fastapi
uvicorn
opencv-python
numpy
deep-translator
httpx
python-multipart
Pillow
gtts
